if: Tidy up convert module and make it public
Move out DMAP related code from convert.py and make what is remaining
part of the public interface.

Fixes #469.
